Fix animation background
The `diagrams-cairo` backend does not have many options for generated
files.  The default background is black, and I am unable to find a way
to set a different default background.  The white rectangle background
in the diagram is aliased, resulting in gray lines at the top and bottom
of the rendered GIF animation.

Unable to fix the issue on the Haskell side, I resorted to using an
external program to crop the rendered GIF file.  Since I am using an
external program anyway, I can also reduce the number of colors in the
images, greatly reducing the image size.